requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.1
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
playwright==1.25.0
pluggy==1.0.0
py==1.11.0
pyee==8.1.0
pyparsing==3.0.9
pytest==7.1.2
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-slugify==6.1.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.11
websockets==10.1
